some refactoring shit + add explore categoryspecific pages
